refactor(node): relocate Node class
