+++ src/node/index.js
@@ -1,0 +1,2 @@
+import Node from './node.js';
+export default Node;
