+++ src/node/modules/children.js
